Revert "Logreg"
