+++ .gitignore
@@ -13,7 +13,3 @@ __pycache__/
 
 # data files
 data/senteval_data*
-
-*hf
-*npy
-*zip
