+++ examples/bert.py
@@ -41,6 +41,7 @@ def batcher(params, batch):
 
 
 # Set params for SentEval
+params_senteval = {'task_path': PATH_TO_DATA, 'usepytorch': True, 'kfold': 5}
 params_senteval['classifier'] = {'nhid': 0, 'optim': 'rmsprop', 'batch_size': 128,
                                  'tenacity': 3, 'epoch_size': 2}
 
@@ -52,6 +53,5 @@ if __name__ == "__main__":
     transfer_tasks = ['Length', 'WordContent', 'Depth', 'TopConstituents',
                         'BigramShift', 'Tense', 'SubjNumber', 'ObjNumber',
                         'OddManOut', 'CoordinationInversion']
-
     results = se.eval(transfer_tasks)
     print(results)
