--- senteval/.ipynb_checkpoints/__init__-checkpoint.py
@@ -1,10 +1,0 @@
-# Copyright (c) 2017-present, Facebook, Inc.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-#
-
-from __future__ import absolute_import
-
-from senteval.engine import SE
