+++ senteval/tools/validation.py
@@ -23,7 +23,6 @@ assert(sklearn.__version__ >= "0.18.0"), \
     "need to update sklearn to version >= 0.18.0"
 from sklearn.linear_model import LogisticRegression
 from sklearn.model_selection import StratifiedKFold
-from catboost import CatBoostClassifier, Pool
 
 
 def get_classif_name(classifier_config, usepytorch):
@@ -102,7 +101,6 @@ class InnerKFoldClassifier(object):
                 clf.fit(X_train, y_train)
 
             self.testresults.append(round(100*clf.score(X_test, y_test), 2))
-           
 
         devaccuracy = round(np.mean(self.devresults), 2)
         testaccuracy = round(np.mean(self.testresults), 2)
@@ -208,7 +206,6 @@ class SplitClassifier(object):
                [2**t for t in range(-2, 4, 1)]
         if self.noreg:
             regs = [1e-9 if self.usepytorch else 1e9]
-            
         scores = []
         for reg in regs:
             if self.usepytorch:
@@ -246,7 +243,68 @@ class SplitClassifier(object):
 
         testaccuracy = clf.score(self.X['test'], self.y['test'])
         testaccuracy = round(100*testaccuracy, 2)
-
         return devaccuracy, testaccuracy
 
+class SplitClassifier2(object):
+    """
+    (train, valid, test) split classifier.
+    """
+    def __init__(self, X, y, config):
+        self.X = X
+        self.y = y
+        self.nclasses = config['nclasses']
+        self.featdim = self.X['train'].shape[1]
+        self.seed = config['seed']
+        self.usepytorch = config['usepytorch']
+        self.classifier_config = config['classifier']
+        self.cudaEfficient = False if 'cudaEfficient' not in config else \
+            config['cudaEfficient']
+        self.modelname = get_classif_name(self.classifier_config, self.usepytorch)
+        self.noreg = False if 'noreg' not in config else config['noreg']
+        self.config = config
+
+    def run(self):
+        logging.info('Training {0} with standard validation..'
+                     .format(self.modelname))
+        regs = [10**t for t in range(-5, -1)] if self.usepytorch else \
+               [2**t for t in range(-2, 4, 1)]
+        if self.noreg:
+            regs = [1e-9 if self.usepytorch else 1e9]
+        scores = []
+        for reg in regs:
+            if self.usepytorch:
+                clf = MLP(self.classifier_config, inputdim=self.featdim,
+                          nclasses=self.nclasses, l2reg=reg,
+                          seed=self.seed, cudaEfficient=self.cudaEfficient)
+
+                # TODO: Find a hack for reducing nb epoches in SNLI
+                clf.fit(self.X['train'], self.y['train'],
+                        validation_data=(self.X['valid'], self.y['valid']))
+            else:
+                clf = LogisticRegression(C=reg, random_state=self.seed)
+                clf.fit(self.X['train'], self.y['train'])
+            scores.append(round(100*clf.score(self.X['valid'],
+                                self.y['valid']), 2))
+        logging.info([('reg:'+str(regs[idx]), scores[idx])
+                      for idx in range(len(scores))])
+        optreg = regs[np.argmax(scores)]
+        devaccuracy = np.max(scores)
+        logging.info('Validation : best param found is reg = {0} with score \
+            {1}'.format(optreg, devaccuracy))
+        clf = LogisticRegression(C=optreg, random_state=self.seed)
+        logging.info('Evaluating...')
+        if self.usepytorch:
+            clf = MLP(self.classifier_config, inputdim=self.featdim,
+                      nclasses=self.nclasses, l2reg=optreg,
+                      seed=self.seed, cudaEfficient=self.cudaEfficient)
+
+            # TODO: Find a hack for reducing nb epoches in SNLI
+            clf.fit(self.X['train'], self.y['train'],
+                    validation_data=(self.X['valid'], self.y['valid']))
+        else:
+            clf = LogisticRegression(C=optreg, random_state=self.seed)
+            clf.fit(self.X['train'], self.y['train'])
+
+        testaccuracy = clf.score(self.X['test'], self.y['test'])
+        testaccuracy = round(100*testaccuracy, 2)
         return clf
